Reset-AppDataFolder: Fixing exception when a "Logs" folder is not present
--HG--
branch : dev

+++ Reset-AppDataFolder/Reset-AppDataFolder.psm1
@@ -47,8 +47,11 @@ function Reset-AppDataFolder
         {
             $whiteList += "Logs"
 
-            # Removing empty files and files that are not log files in the "Logs" folder.
-            Get-ChildItem -Path "$Path\Logs" | Where-Object { $PSItem.Extension -ne ".log" -or $PSItem.Length -eq 0 } | Remove-Item -Force
+            if (Test-Path("$Path\Logs"))
+            {
+                # Removing empty files and files that are not log files in the "Logs" folder.
+                Get-ChildItem -Path "$Path\Logs" | Where-Object { $PSItem.Extension -ne ".log" -or $PSItem.Length -eq 0 } | Remove-Item -Force                
+            }
         }
 
         Get-ChildItem -Path $Path | Where-Object { $whiteList -notcontains $PSItem.Name } | Remove-Item -Recurse -Force
